Update galevo_nitrogen.py
Added gas infall model: added a new input parameter, tau_infalle9, for the galaxy_evol() function with a default value of 0. tau_infalle9=0 is equivalent to the previous code without gas infall. Set the value of tau_infalle9, which is the gas infall timescale in the unit of [Gyr], to activate the gas infall model.